pyproject.toml
[project]
name = "speech-decoding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "einops>=0.8.1",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "jieba>=0.42.1",
    "matplotlib>=3.10.8",
    "nltk>=3.9.2",
    "numpy>=2.4.0",
    "omegaconf>=2.3.0",
    "opencc-python-reimplemented>=0.1.7",
    "pandas>=2.3.3",
    "psutil>=7.2.0",
    "pypinyin>=0.55.0",
    "python-box>=7.3.2",
    "python-levenshtein>=0.27.3",
    "regex>=2025.11.3",
    "rouge-chinese>=1.0.3",
    "scikit-learn>=1.8.0",
    "speechbrain>=1.0.3",
    "timm>=1.0.22",
    "torch==2.4.1",
    "torchaudio==2.4.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.3",
    "pinyin2hanzi>=0.1.0",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"
extra-index-url = [
    "https://mirrors.aliyun.com/pypi/simple"
]
